Basic pickup item + Player health
Gave the player health and created a pickup health item with a pickup base file which can be used for any future pickup items as well.  The pickup health is in the middle of the hallway right now for testing but just needs to be dragged into prefabs folder to be made a prefab, will be working on player inventory next.