Attempting to push hotbar and gun shooter changes
Git is being annoying but I have the hotbar function working and your keyboard 1-9 keys are the slot numbers when you hit the number on your keyboard it uses that item. ItemDictionary keeps track of items and their id numbers, gun pickup automatically equips to user and shoots where you click. Bullets need to be slightly fixed still and the gun moves while on the ground too so that need to be fixed, but mostly working now.